requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0
python-dotenv==1.0.1
